requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
python-dotenv>=1.0.1
aiohttp>=3.9.0
boto3>=1.34.0
opencv-contrib-python>=4.8.1.78
numpy>=1.24.0
scikit-image>=0.22.0
scipy>=1.10.0
pillow>=10.0.0
pydantic>=2.7.0
